+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/MessageListView.kt
@@ -53,13 +53,26 @@ import io.getstream.chat.android.client.models.Reaction
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.utils.Result
 import io.getstream.chat.android.common.model.ModeratedMessageOption
+import io.getstream.chat.android.common.state.Copy
+import io.getstream.chat.android.common.state.CustomAction
+import io.getstream.chat.android.common.state.Delete
 import io.getstream.chat.android.common.state.DeletedMessageVisibility
+import io.getstream.chat.android.common.state.Edit
+import io.getstream.chat.android.common.state.MessageAction
+import io.getstream.chat.android.common.state.MuteUser
+import io.getstream.chat.android.common.state.Pin
+import io.getstream.chat.android.common.state.React
+import io.getstream.chat.android.common.state.Reply
+import io.getstream.chat.android.common.state.Resend
+import io.getstream.chat.android.common.state.ThreadReply
 import io.getstream.chat.android.core.internal.InternalStreamChatApi
 import io.getstream.chat.android.core.internal.coroutines.DispatcherProvider
 import io.getstream.chat.android.offline.extensions.downloadAttachment
+import io.getstream.chat.android.offline.extensions.globalState
 import io.getstream.chat.android.ui.ChatUI
 import io.getstream.chat.android.ui.R
 import io.getstream.chat.android.ui.common.extensions.getCreatedAtOrThrow
+import io.getstream.chat.android.ui.common.extensions.internal.copyToClipboard
 import io.getstream.chat.android.ui.common.extensions.internal.createStreamThemeWrapper
 import io.getstream.chat.android.ui.common.extensions.internal.getFragmentManager
 import io.getstream.chat.android.ui.common.extensions.internal.isCurrentUser
@@ -84,6 +97,7 @@ import io.getstream.chat.android.ui.message.list.MessageListView.AttachmentDownl
 import io.getstream.chat.android.ui.message.list.MessageListView.BottomEndRegionReachedHandler
 import io.getstream.chat.android.ui.message.list.MessageListView.ConfirmDeleteMessageHandler
 import io.getstream.chat.android.ui.message.list.MessageListView.ConfirmFlagMessageHandler
+import io.getstream.chat.android.ui.message.list.MessageListView.CustomActionHandler
 import io.getstream.chat.android.ui.message.list.MessageListView.EndRegionReachedHandler
 import io.getstream.chat.android.ui.message.list.MessageListView.EnterThreadListener
 import io.getstream.chat.android.ui.message.list.MessageListView.ErrorEventHandler
@@ -122,12 +136,13 @@ import io.getstream.chat.android.ui.message.list.background.MessageBackgroundFac
 import io.getstream.chat.android.ui.message.list.background.MessageBackgroundFactoryImpl
 import io.getstream.chat.android.ui.message.list.internal.HiddenMessageListItemPredicate
 import io.getstream.chat.android.ui.message.list.internal.MessageListScrollHelper
+import io.getstream.chat.android.ui.message.list.options.message.MessageOptionItemsFactory
 import io.getstream.chat.android.ui.message.list.options.message.internal.MessageOptionsDialogFragment
-import io.getstream.chat.android.ui.message.list.options.message.internal.MessageOptionsView
 import io.getstream.chat.android.ui.utils.extensions.isCurrentUserBanned
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
+import io.getstream.chat.android.common.state.Flag as FlagAction
 
 /**
  * MessageListView renders a list of messages and extends the [RecyclerView]
@@ -220,6 +235,9 @@ public class MessageListView : ConstraintLayout {
     }
     private var userBlockHandler = UserBlockHandler { _, _ ->
         throw IllegalStateException("onBlockUserHandler must be set.")
+    }
+    private var customActionHandler = CustomActionHandler { _, _ ->
+        throw IllegalStateException("onCustomActionHandler must be set.")
     }
     private var messageReplyHandler = MessageReplyHandler { _, _ ->
         throw IllegalStateException("onReplyMessageHandler must be set")
@@ -349,50 +367,32 @@ public class MessageListView : ConstraintLayout {
                     val isEditEnabled = style.editMessageEnabled && !message.isGiphyNotEphemeral()
                     val viewStyle = style.copy(editMessageEnabled = isEditEnabled)
 
+                    val messageOptionItems = messageOptionItemsFactory.createMessageOptionItems(
+                        selectedMessage = message,
+                        currentUser = ChatUI.currentUserProvider.getCurrentUser(),
+                        isInThread = adapter.isThread || message.isInThread(),
+                        ownCapabilities = ownCapabilities,
+                        style = viewStyle
+                    )
+
                     MessageOptionsDialogFragment
                         .newMessageOptionsInstance(
-                            message,
-                            MessageOptionsView.Configuration(
-                                viewStyle = viewStyle,
-                                channelConfig = channel.config,
-                                hasTextToCopy = message.text.isNotBlank(),
-                                suppressThreads = adapter.isThread || message.isInThread(),
-                                ownCapabilities = ownCapabilities
-                            ),
-                            viewStyle,
-                            messageListItemViewHolderFactory,
-                            messageBackgroundFactory,
-                            attachmentFactoryManager,
-                            showAvatarPredicate
+                            message = message,
+                            style = viewStyle,
+                            messageViewHolderFactory = messageListItemViewHolderFactory,
+                            messageBackgroundFactory = messageBackgroundFactory,
+                            attachmentFactoryManager = attachmentFactoryManager,
+                            showAvatarPredicate = showAvatarPredicate,
+                            messageOptionItems = messageOptionItems,
                         )
                         .apply {
                             setReactionClickHandler { message, reactionType ->
                                 messageReactionHandler.onMessageReaction(message, reactionType)
                             }
-                            setConfirmDeleteMessageClickHandler { message, callback ->
-                                confirmDeleteMessageHandler.onConfirmDeleteMessage(
-                                    message,
-                                    callback::onConfirmDeleteMessage
-                                )
+
+                            setMessageActionClickHandler { messageAction ->
+                                handleMessageAction(messageAction)
                             }
-                            setConfirmFlagMessageClickHandler { message, callback ->
-                                confirmFlagMessageHandler.onConfirmFlagMessage(message, callback)
-                            }
-                            setMessageOptionsHandlers(
-                                MessageOptionsDialogFragment.MessageOptionsHandlers(
-                                    threadReplyHandler = threadStartHandler,
-                                    retryHandler = messageRetryHandler,
-                                    editClickHandler = messageEditHandler,
-                                    flagClickHandler = messageFlagHandler,
-                                    pinClickHandler = messagePinHandler,
-                                    unpinClickHandler = messageUnpinHandler,
-                                    muteClickHandler = userMuteHandler,
-                                    unmuteClickHandler = userUnmuteHandler,
-                                    blockClickHandler = userBlockHandler,
-                                    deleteClickHandler = messageDeleteHandler,
-                                    replyClickHandler = messageReplyHandler,
-                                )
-                            )
                         }
                         .show(fragmentManager, MessageOptionsDialogFragment.TAG)
                 }
@@ -498,13 +498,6 @@ public class MessageListView : ConstraintLayout {
             context.getFragmentManager()?.let {
                 MessageOptionsDialogFragment.newReactionOptionsInstance(
                     message,
-                    MessageOptionsView.Configuration(
-                        viewStyle = requireStyle(),
-                        channelConfig = channel.config,
-                        hasTextToCopy = false, // No effect when displaying reactions
-                        suppressThreads = false, // No effect when displaying reactions
-                        ownCapabilities = ownCapabilities
-                    ),
                     requireStyle(),
                     messageListItemViewHolderFactory,
                     messageBackgroundFactory,
@@ -555,6 +548,7 @@ public class MessageListView : ConstraintLayout {
     private lateinit var messageDateFormatter: DateFormatter
     private lateinit var attachmentFactoryManager: AttachmentFactoryManager
     private lateinit var messageBackgroundFactory: MessageBackgroundFactory
+    private lateinit var messageOptionItemsFactory: MessageOptionItemsFactory
 
     public constructor(context: Context) : this(context, null, 0)
     public constructor(context: Context, attrs: AttributeSet?) : this(context, attrs, 0)
@@ -771,6 +765,10 @@ public class MessageListView : ConstraintLayout {
 
         if (::messageBackgroundFactory.isInitialized.not()) {
             messageBackgroundFactory = MessageBackgroundFactoryImpl(requireStyle().itemStyle)
+        }
+
+        if (::messageOptionItemsFactory.isInitialized.not()) {
+            messageOptionItemsFactory = MessageOptionItemsFactory.defaultFactory(context)
         }
 
         messageListItemViewHolderFactory.decoratorProvider = MessageListItemDecoratorProvider(
@@ -1010,6 +1008,19 @@ public class MessageListView : ConstraintLayout {
             "Adapter was already initialized, please set MessageBackgroundFactory first"
         }
         this.messageBackgroundFactory = messageBackgroundFactory
+    }
+
+    /**
+     * Allows clients to set a custom implementation of [MessageOptionItemsFactory]. Use this
+     * method if you want to change the message options on the message options overlay.
+     *
+     * @param messageOptionItemsFactory The custom factory that provides option items for the message options overlay.
+     */
+    public fun setMessageOptionItemsFactory(messageOptionItemsFactory: MessageOptionItemsFactory) {
+        check(isAdapterInitialized().not()) {
+            "Adapter was already initialized, please set MessageOptionItemsFactory first"
+        }
+        this.messageOptionItemsFactory = messageOptionItemsFactory
     }
 
     /**
@@ -1438,8 +1449,23 @@ public class MessageListView : ConstraintLayout {
      *
      * @param userBlockHandler The handler to use.
      */
+    @Deprecated(
+        message = "The block action has been removed. Use MessageOptionItemsFactory.setMessageOptionItemsFactory() " +
+            "in conjunction with MessageOptionItemsFactory.setCustomActionHandler() to add support for custom block " +
+            "action.",
+        level = DeprecationLevel.ERROR
+    )
     public fun setUserBlockHandler(userBlockHandler: UserBlockHandler) {
         this.userBlockHandler = userBlockHandler
+    }
+
+    /**
+     * Set the handler used when the custom action is going to be executed.
+     *
+     * @param customActionHandler The handler to use.
+     */
+    public fun setCustomActionHandler(customActionHandler: CustomActionHandler) {
+        this.customActionHandler = customActionHandler
     }
 
     /**
@@ -1550,6 +1576,61 @@ public class MessageListView : ConstraintLayout {
      */
     public fun setModeratedMessageHandler(handler: ModeratedMessageOptionHandler) {
         this.moderatedMessageOptionHandler = handler
+    }
+
+    /**
+     * Handles the selected [messageAction].
+     *
+     * @param messageAction The newly selected action.
+     */
+    private fun handleMessageAction(messageAction: MessageAction) {
+        val message = messageAction.message
+        val style = requireStyle()
+
+        when (messageAction) {
+            is Resend -> messageRetryHandler.onMessageRetry(message)
+            is Reply -> messageReplyHandler.onMessageReply(message.cid, message)
+            is ThreadReply -> threadStartHandler.onStartThread(message)
+            is Copy -> context.copyToClipboard(message.text)
+            is Edit -> messageEditHandler.onMessageEdit(message)
+            is Pin -> {
+                if (message.pinned) {
+                    messageUnpinHandler.onMessageUnpin(message)
+                } else {
+                    messagePinHandler.onMessagePin(message)
+                }
+            }
+            is Delete -> {
+                if (style.deleteConfirmationEnabled) {
+                    confirmDeleteMessageHandler.onConfirmDeleteMessage(message) {
+                        messageDeleteHandler.onMessageDelete(message)
+                    }
+                } else {
+                    messageDeleteHandler.onMessageDelete(message)
+                }
+            }
+            is FlagAction -> {
+                if (style.flagMessageConfirmationEnabled) {
+                    confirmFlagMessageHandler.onConfirmFlagMessage(message) {
+                        messageFlagHandler.onMessageFlag(message)
+                    }
+                } else {
+                    messageFlagHandler.onMessageFlag(message)
+                }
+            }
+            is MuteUser -> {
+                val isUserMuted = ChatClient.instance().globalState.muted.value.any { it.target.id == message.user.id }
+                if (isUserMuted) {
+                    userUnmuteHandler.onUserUnmute(message.user)
+                } else {
+                    userMuteHandler.onUserMute(message.user)
+                }
+            }
+            is CustomAction -> customActionHandler.onCustomAction(message, messageAction.extraProperties)
+            is React -> {
+                // Handled by a separate handler.
+            }
+        }
     }
     //endregion
 
@@ -1696,6 +1777,10 @@ public class MessageListView : ConstraintLayout {
 
     public fun interface UserBlockHandler {
         public fun onUserBlock(user: User, cid: String)
+    }
+
+    public fun interface CustomActionHandler {
+        public fun onCustomAction(message: Message, extraProperties: Map<String, Any>)
     }
 
     public fun interface AttachmentDownloadHandler {